% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-descriptives.R
\name{build_ordinal_type}
\alias{build_ordinal_type}
\title{Build an ordinal type.}
\usage{
build_ordinal_type(df_column)
}
\arguments{
\item{df_column}{A vector such as the column of a dataframe}
}
\value{
A vector as an ordered factor
}
\description{
This function returns a column of a dataframe as an ordered factor (ordinal)
}
\examples{
build_ordinal_type(mtcars$mpg)
}
