% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-descriptives.R
\name{generateDescriptives}
\alias{generateDescriptives}
\title{Generates descriptive table for the specified list of dataframes.}
\usage{
generateDescriptives(
  listDataframes,
  cohortCol = NULL,
  cohortNames = NULL,
  useCols = NULL,
  excludeCols = NULL,
  column_types = NULL,
  continousStatAgg = "both",
  continousTest = "auto",
  categoricalTest = "auto",
  dig = 2,
  csv = F
)
}
\arguments{
\item{listDataframes}{a single dataframe or a list of dataframes}

\item{cohortCol}{character vector indicating the column name of listDataframes to
be used for cohort assignment}

\item{cohortNames}{character vector indicating the names of the cohorts}

\item{useCols}{character vector indicating the names of the columns to be used}

\item{excludeCols}{character vector indicating the names of the columns to be excluded}

\item{column_types}{List indicating the names of the columns and the type of the variables}

\item{continousTest}{character indicating the test for the continous variables
possible values: "auto", "anova" , "wt".}

\item{categoricalTest}{character indicating the test for the categorical variables
possible values: "auto", "chisq" , "fe".}

\item{dig}{Integer indicating the number of decimal places to be displayed.}

\item{csv}{csv: Boolean, indicating if the output should be exported as a CSV. By default
the output is not exported.}

\item{continuousStatAgg}{character indicating in the continous variables
to be summarized with the Mean (SD), Median (IQR), or both in consecutive rows.
Possible values: "mean", "median", "both"}
}
\value{
A table with the described statistics
}
\description{
This function is used to obtain descriptive statistics for a given list of dataframes
or patient ids. The summary table created can be viewed as a markdown table on the stdout or optionally export to csv.
}
\examples{
library(NeuroBlu)
library(survival)
generateDescriptives(diabetic, cohortCol = "trt", useCols = c("age", "eye", "risk", "time", "trt"), csv = T)

For ordinal variables:

test <- subset(mockstudy, select = c(sex, age))
age_ordinal <- build_ordinal_type(test$age)
list_types <- list("age" = age_ordinal)
generateDescriptives(test, cohortCol = "sex",column_types = list_types)
}
