% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptives.R
\name{get_descriptives}
\alias{get_descriptives}
\title{Function to obtain descriptive statistics of a given list of dataframes
or patients.}
\usage{
get_descriptives(
  list_dataframes,
  cohort_col = NULL,
  cohort_names = NULL,
  use_cols = NULL,
  exclude_cols = NULL,
  continous_stat_agg = "both",
  continous_test = "auto",
  categorical_test = "auto",
  dig = 2,
  csv = F
)
}
\arguments{
\item{list_dataframes}{a single dataframe or a list of dataframes}

\item{cohort_col}{character vector indicating the colname of list_dataframes to
be used for cohort assignment}

\item{cohort_names}{character vector indicating the names of the cohorts}

\item{use_cols}{character vector indicating the names of the columns to be used}

\item{exclude_cols}{character vector indicating the names of the columns to be excluded}

\item{continous_test}{character indicating the test for the continous variables
possible values: "auto", "anova" , "wt".}

\item{categorical_test}{character indicating the test for the continous variables
possible values: "auto", "chisq" , "fe".}

\item{dig}{Integer indicating the number of decimal places to be shown.}

\item{csv}{csv: Boolean, indicating if the output should be a CSV. By default
the output is in Markdown}

\item{continuous_stat_agg}{haracter indicating in the continous variables
should be summarized with the Mean (STD), Median (IQR), or both in 2 rows.
Possible values: "mean", "median", "both"}
}
\value{
A table with the described statistics
}
\description{
Function to obtain descriptive statistics of a given list of dataframes
or patients.
}
\examples{
get_descriptives(mtcars)
get_descriptives(list(list(mtcars[1:10,]), mtcars[11:20,]))
}
